pyproject.toml
[project]
name = "micro-agent-protocol"
version = "0.1.0"
description = "A 50-line YAML that runs anywhere - Define agent workflows in a single file"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Robert Song", email = "robertsong@example.com"}
]
keywords = ["ai", "agent", "workflow", "yaml", "automation", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.scripts]
map = "map_cli:main"

[project.urls]
Homepage = "https://github.com/robertsong2019/micro-agent-protocol"
Documentation = "https://github.com/robertsong2019/micro-agent-protocol#readme"
Repository = "https://github.com/robertsong2019/micro-agent-protocol"
Issues = "https://github.com/robertsong2019/micro-agent-protocol/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*.py", "compilers/*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
